Fix auto-sync interval selection
The auto-sync interval selection was not displaying the selected value correctly. This commit addresses the issue by ensuring that the selected interval is properly reflected in the UI.

+++ src/components/dashboard/EmailForwarding.tsx
@@ -593,7 +593,7 @@ export function EmailForwarding() {
                                 updateAccountSyncMutation.mutate({
                                   accountId: account.id,
                                   autoSyncEnabled: true,
-                                  syncIntervalMinutes: parseInt(value)
+                                  syncIntervalMinutes: parseFloat(value)
                                 })
                               }
                               disabled={updateAccountSyncMutation.isPending}
